Axon Segmentation/Axon segmentation model training/requirements.txt
axial_attention==0.6.1
batchgenerators==0.23
dicom2nifti==2.3.0
einops==0.3.2
h5py==3.1.0
hiddenlayer==0.3
imgaug==0.4.0
matplotlib==3.4.2
MedPy==0.4.0
nibabel==3.2.2
numpy==1.20.3
numpy==1.19.5
opencv_python==4.5.3.56
pandas==1.3.2
Pillow==8.3.1
Pillow==9.5.0
Requests==2.29.0
scikit_image==0.18.3
scikit_learn==1.0.2
scipy==1.7.0
setuptools==52.0.0.post20210125
SimpleITK==2.1.0
SimpleITK==2.2.1
tifffile==2020.10.1
torch==1.9.0
tqdm==4.61.2
unittest2==1.1.0
